Fix project-specific information (prompt, encoding, etc.) not being provided when the project is activated at startup (ActivateProjectTool not applied)
Peripheral change: Move MemoriesManager into module 'project',
moving the instance into the Project instance, because memories are strictly
associated with the project

Fixes #681
