+++ src/serena/agent.py
@@ -11,7 +11,6 @@ import webbrowser
 from collections.abc import Callable
 from concurrent.futures import Future, ThreadPoolExecutor
 from logging import Logger
-from pathlib import Path
 from typing import TYPE_CHECKING, Any, Optional, TypeVar
 
 from sensai.util import logging
@@ -21,8 +20,7 @@ from interprompt.jinja_template import JinjaTemplate
 from serena import serena_version
 from serena.analytics import RegisteredTokenCountEstimator, ToolUsageStats
 from serena.config.context_mode import RegisteredContext, SerenaAgentContext, SerenaAgentMode
-from serena.config.serena_config import SerenaConfig, ToolInclusionDefinition, ToolSet, get_serena_managed_in_project_dir
-from serena.constants import SERENA_FILE_ENCODING
+from serena.config.serena_config import SerenaConfig, ToolInclusionDefinition, ToolSet
 from serena.dashboard import SerenaDashboardAPI
 from serena.project import Project
 from serena.prompt_factory import SerenaPromptFactory
@@ -42,40 +40,6 @@ SUCCESS_RESULT = "OK"
 
 class ProjectNotFoundError(Exception):
     pass
-
-
-class MemoriesManager:
-    def __init__(self, project_root: str):
-        self._memory_dir = Path(get_serena_managed_in_project_dir(project_root)) / "memories"
-        self._memory_dir.mkdir(parents=True, exist_ok=True)
-        self._encoding = SERENA_FILE_ENCODING
-
-    def _get_memory_file_path(self, name: str) -> Path:
-        # strip all .md from the name. Models tend to get confused, sometimes passing the .md extension and sometimes not.
-        name = name.replace(".md", "")
-        filename = f"{name}.md"
-        return self._memory_dir / filename
-
-    def load_memory(self, name: str) -> str:
-        memory_file_path = self._get_memory_file_path(name)
-        if not memory_file_path.exists():
-            return f"Memory file {name} not found, consider creating it with the `write_memory` tool if you need it."
-        with open(memory_file_path, encoding=self._encoding) as f:
-            return f.read()
-
-    def save_memory(self, name: str, content: str) -> str:
-        memory_file_path = self._get_memory_file_path(name)
-        with open(memory_file_path, "w", encoding=self._encoding) as f:
-            f.write(content)
-        return f"Memory {name} written."
-
-    def list_memories(self) -> list[str]:
-        return [f.name.replace(".md", "") for f in self._memory_dir.iterdir() if f.is_file()]
-
-    def delete_memory(self, name: str) -> str:
-        memory_file_path = self._get_memory_file_path(name)
-        memory_file_path.unlink()
-        return f"Memory {name} deleted."
 
 
 class AvailableTools:
@@ -123,7 +87,6 @@ class SerenaAgent:
         # project-specific instances, which will be initialized upon project activation
         self._active_project: Project | None = None
         self.language_server: SolidLanguageServer | None = None
-        self.memories_manager: MemoriesManager | None = None
 
         # adjust log level
         serena_log_level = self.serena_config.log_level
@@ -361,6 +324,11 @@ class SerenaAgent:
             available_tools=self._exposed_tools.tool_names,
             available_markers=available_markers,
         )
+
+        # If a project is active at startup, append its activation message
+        if self._active_project is not None:
+            system_prompt += "\n\n" + self._active_project.get_activation_message()
+
         log.info("System prompt:\n%s", system_prompt)
         return system_prompt
 
@@ -425,9 +393,6 @@ class SerenaAgent:
         log.info(f"Activating {project.project_name} at {project.project_root}")
         self._active_project = project
         self._update_active_tools()
-
-        # initialize project-specific instances which do not depend on the language server
-        self.memories_manager = MemoriesManager(project.project_root)
 
         def init_language_server() -> None:
             # start the language server
